feat: add image to post share.
